+++ src/anyio/abc/streams.py
@@ -89,6 +89,15 @@ class ObjectStream(Generic[T_Item], ObjectReceiveStream[T_Item], ObjectSendStrea
     A bidirectional message stream which guarantees the order and reliability of message delivery.
     """
 
+    @abstractmethod
+    async def send_eof(self) -> None:
+        """
+        Send an end-of-file indication to the peer.
+
+        You should not try to send any further data to this stream after calling this method.
+        This method is idempotent (does nothing on successive calls).
+        """
+
 
 class ByteReceiveStream(AsyncResource):
     """
@@ -132,6 +141,15 @@ class ByteSendStream(AsyncResource):
 
 class ByteStream(ByteReceiveStream, ByteSendStream):
     """A bidirectional byte stream."""
+
+    @abstractmethod
+    async def send_eof(self) -> None:
+        """
+        Send an end-of-file indication to the peer.
+
+        You should not try to send any further data to this stream after calling this method.
+        This method is idempotent (does nothing on successive calls).
+        """
 
 
 #: type alias for all unreliable bytes-oriented receive streams